feat: Create core:network module and migrate network infrastructure
Extracted Retrofit configuration, OkHttp providers, and network error handling logic from the data module into a dedicated core:network library. This improves modularity and separates base network infrastructure from data-specific implementations, using a new NetworkConfig interface to handle app-level configurations.
