+++ core/network/src/main/java/com/tyme/github/users/core/network/config/NetworkConfig.kt
@@ -1,0 +1,8 @@
+package com.tyme.github.users.core.network.config
+
+interface NetworkConfig {
+    val baseUrl: String
+    val baseDomain: String
+    val pinningPublicKey: String
+    val isDebug: Boolean
+}
