+++ core/network/src/main/java/com/tyme/github/users/core/network/responses/errors/ErrorResponse.kt
@@ -1,0 +1,10 @@
+package com.tyme.github.users.core.network.responses.errors
+
+import com.google.gson.annotations.SerializedName
+
+data class ErrorResponse(
+    @SerializedName("code")
+    val code: Int? = null,
+    @SerializedName("message")
+    val message: String? = null,
+)
