+++ data/src/main/java/com/tyme/github/users/data/di/modules/ProvidersModule.kt
@@ -1,11 +1,7 @@
 package com.tyme.github.users.data.di.modules
 
-import com.tyme.github.users.data.providers.CertificatePinnersProvider
-import com.tyme.github.users.data.providers.CertificatePinnersProviderImpl
-import com.tyme.github.users.data.providers.OkHttpClientsProvider
-import com.tyme.github.users.data.providers.OkHttpClientsProviderImpl
-import com.tyme.github.users.data.providers.RetrofitsProvider
-import com.tyme.github.users.data.providers.RetrofitsProviderImpl
+import com.tyme.github.users.core.network.config.NetworkConfig
+import com.tyme.github.users.data.network.NetworkConfigImpl
 import com.tyme.github.users.data.providers.SecretKeysProvider
 import com.tyme.github.users.data.providers.SecretKeysProviderImpl
 import dagger.Binds
@@ -21,11 +17,5 @@ internal interface ProvidersModule {
     fun bindKeysProvider(impl: SecretKeysProviderImpl): SecretKeysProvider
 
     @Binds
-    fun bindCertificatePinnersProvider(impl: CertificatePinnersProviderImpl): CertificatePinnersProvider
-
-    @Binds
-    fun bindOkHttpClientsProvider(impl: OkHttpClientsProviderImpl): OkHttpClientsProvider
-
-    @Binds
-    fun bindRetrofitsProvider(impl: RetrofitsProviderImpl): RetrofitsProvider
+    fun bindNetworkConfig(impl: NetworkConfigImpl): NetworkConfig
 }
