--- data/src/main/java/com/tyme/github/users/data/remote/responses/errors/ErrorResponse.kt
@@ -1,10 +1,0 @@
-package com.tyme.github.users.data.remote.responses.errors
-
-import com.google.gson.annotations.SerializedName
-
-data class ErrorResponse(
-    @SerializedName("code")
-    val code: Int? = null,
-    @SerializedName("message")
-    val message: String? = null,
-)
